큐레이션 작성 페이지 리팩토링
