+++ Kindy/Kindy/Network Controllers/CurationRequest.swift
@@ -56,4 +56,17 @@ extension CurationRequest {
             }
         }
     }
+    
+    func asyncUploadImage(image: UIImage, pathRoot: String) async throws -> String? {
+        guard let imageData = image.jpegData(compressionQuality: 0.1) else { return nil }
+        let metaData = StorageMetadata()
+        metaData.contentType = "image/jpeg"
+        let imageName = UUID().uuidString + String(Date().timeIntervalSince1970)
+        
+        let firebaseReference = Storage.storage().reference().child("\(pathRoot)/\(imageName)")
+        let _ = try await firebaseReference.putDataAsync(imageData, metadata: metaData)
+        let url = try await firebaseReference.downloadURL().absoluteString
+        return url
+    }
+    
 }
